requirements.txt
numpy==1.26.2
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
pymysql==1.1.0
Flask-Migrate==4.0.5
python_dotenv==1.0.0